+++ apps/ui/src/routes/tasks/[id]/+page.svelte
@@ -153,7 +153,7 @@
     {#if task}
         <TaskPage
             {task}
-            {runs}
+            bind:runs
             {concurrencyReached}
             {triggering}
             {stopping}